jump fix/ spawner fix/ merge/ scorehandle update
Heb wat scripts opgeschoont en wat verbetert. Ook de jump is veranderd.
verder is de scene gemerged en zit alles nu in elkaar.

+++ UnityProject/Assets/Scripts/HUDScript.cs
@@ -9,7 +9,7 @@ public class HUDScript : MonoBehaviour {
 		playerScore += Time.deltaTime;
 	}
 	
-	public void IncreaseScore(int amount)
+	public void IncreaseScore(float amount)
 	{ 
 		playerScore += amount;
 	}
